Fixed a problem with the way constants where handled in SSA.
By default constants where not pushed again to SSA if they already
had a location. But this method causes problems with aliased values.
Aliased values assume that the either one of the values is loaded in
a block of SSA instructions.
but this can not be the case with constants.